==> PGM/08_model_improvements.py <==
# -*- coding: utf-8 -*-
"""
Created on Thu Feb 28 18:30:38 2019

@author: Chris
"""

# Notes:
# De-weight value of observations for dry period? They're getting equal weight, which seems off
# Alternatively, downsample less?


# Also: Check how a fitted model does on test data. 

# Also also: Correlation should maybe be stronger? Weaker?  Different, definitely. Maybe too strong of correlation is
